refactor: create screen
